Minor auto refresh cleanup
